move into image manager
